Fix Shift/Reduce conflict with VarDecl* and Stmt*
There was a shift/reduce conflict that would arise when the parser reached a
state that looked like this:

        VarDecl*:      VarDecl* . VarDecl
        StmtBlock: '{' VarDecl* . Stmt* '}'

Since the VarDecl* and Stmt* productions could both expand to epsilon (the
empty string), when a T_Identifier token is seen,  the parser can't tell
whether it should reduce or shift the T_Identifier token and try to match
something bigger.

To fix this, the VarDecl* and Stmt* productions were converted to VarDecl+ and
Stmt+ productions respectively. Neither of these new productions can expand to
epsilon (the empty string). Thus, the parser no longer gets confused when
trying to reduce these productions using the empty string.

Also, the StmtBlock production also had to be modified to accommodate these
changes.