Add ability to pass arguments to check.sh
Modified the check.sh script to allow the user to pass in the names of the
specific tests to be run as arguments to the script. Thus, if a user didn't
want to test against all of the test, and perhaps only wanted to test against
the 'simple' test (i.e. simple.out and simple.[decaf|frag]), this could be done
like so:

        ./check.sh simple

Multiple test names are accepted. For example, the following is valid:

        ./check.sh simple functions class