Modify ReturnStmt use location of T_Return token
The ReturnStmt production was previously using the location of it's optional
expression part. Since the expression is not guaranteed to exist, it is better
to use the location of a mandatory token, such as the T_Return token. This bug
was caught by the samples/control.decaf test, and this change should allow this
test to pass.